* move Project Tree stuff from libavogadro to avogadro * implement a QAbstractItemModel for the tree * lots of bug fixes
git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1632 af78ad9f-9711-0410-ba57-a8eff4914ec4

+++ libavogadro/src/cube.cpp
@@ -119,7 +119,7 @@ namespace Avogadro {
 
   bool Cube::setData(const std::vector<double> &values)
   {
-    if (values.size() == m_points.x() * m_points.y() * m_points.z()) {
+    if (static_cast<int>(values.size()) == m_points.x() * m_points.y() * m_points.z()) {
       m_data = values;
       qDebug() << "Loaded in cube data" << m_data.size();
       return true;
